Bugfixed cmd now that they are combined
